break out locate tests
